hopefully no exposed keys

+++ scripts/build-lambdas.sh
@@ -4,6 +4,9 @@ for i in "./lambdas/"*
 do
 	pkg=`basename "$i"`
 	echo "Building $pkg"
-	pip3 install --target "./build/lambda.$pkg" -r "$i/requirements.txt"
-	cp -rp "$i/"* "./build/lambda.$pkg"
+	mkdir -p "./build/lambda.$pkg"
+	if [ -f "$i/requirements.txt" ]; then
+		pip3 install --target "./build/lambda.$pkg" -r "$i/requirements.txt"
+	fi
+	cp -rp "$i/"* "./build/lambda.$pkg/"
 done
